Update scenario grammar
1.
Only one MLM, the MLM in which the validation category is located,
should be tested and executed. Other MLMs are only allowed to execute
when called in the logic slot of the main MLM. 

The following calls the main MLM and checks if it is triggered:
WHEN THE MLM IS CALLED;
THEN THE MLM IS TRIGGERED;
This check is not possible for other MLMs.

2.
Arguments and delay of calls can be checked with indeterminate
expressions:
THEN MLM 'other_mlm' IS CALLED WITH ANYTHING WHERE IT IS > 5, "xyz", 123
DELAY ANYTHING WHERE IT IS WITHIN 5 SECONDS TO 1 MINUTE;